compute index in worker: less lags
